+++ truffle/src/com.oracle.truffle.api.debug/src/com/oracle/truffle/api/debug/DebuggerSession.java
@@ -55,6 +55,7 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Predicate;
+import java.util.function.Supplier;
 
 import com.oracle.truffle.api.Assumption;
 import com.oracle.truffle.api.CompilerAsserts;
@@ -80,7 +81,6 @@ import com.oracle.truffle.api.instrumentation.ProbeNode;
 import com.oracle.truffle.api.instrumentation.SourceSectionFilter;
 import com.oracle.truffle.api.instrumentation.TruffleInstrument;
 import com.oracle.truffle.api.instrumentation.SourceSectionFilter.Builder;
-import com.oracle.truffle.api.instrumentation.StandardTags.CallTag;
 import com.oracle.truffle.api.instrumentation.StandardTags.RootTag;
 import com.oracle.truffle.api.nodes.ExecutableNode;
 import com.oracle.truffle.api.nodes.LanguageInfo;
@@ -187,11 +187,10 @@ public final class DebuggerSession implements Closeable {
     private final SuspendedCallback callback;
     private final Set<SourceElement> sourceElements;
     private final boolean hasExpressionElement;
+    private final boolean hasRootElement;
     private final List<Breakpoint> breakpoints = Collections.synchronizedList(new ArrayList<>());
 
-    private EventBinding<? extends ExecutionEventNodeFactory> callBinding;
     private EventBinding<? extends ExecutionEventNodeFactory> syntaxElementsBinding;
-    private EventBinding<? extends ExecutionEventNodeFactory> rootBinding;
     final Set<EventBinding<? extends ExecutionEventNodeFactory>> allBindings = Collections.synchronizedSet(new HashSet<>());
 
     private final ConcurrentHashMap<Thread, SuspendedEvent> currentSuspendedEventMap = new ConcurrentHashMap<>();
@@ -228,6 +227,7 @@ public final class DebuggerSession implements Closeable {
                 break;
         }
         this.hasExpressionElement = this.sourceElements.contains(SourceElement.EXPRESSION);
+        this.hasRootElement = this.sourceElements.contains(SourceElement.ROOT);
         if (Debugger.TRACE) {
             trace("open with callback %s", callback);
         }
@@ -490,35 +490,27 @@ public final class DebuggerSession implements Closeable {
 
     private void addBindings(boolean includeInternalCode, Predicate<Source> sFilter) {
         if (syntaxElementsBinding == null && !sourceElements.isEmpty()) {
-            // The order of registered instrumentations matters.
-            // It's important to instrument root nodes first to intercept stack changes,
-            // then instrument statements, and
-            // call bindings need to be called after statements.
-            this.rootBinding = createBinding(includeInternalCode, sFilter, new ExecutionEventNodeFactory() {
-                @Override
-                public ExecutionEventNode create(EventContext context) {
-                    return new RootSteppingDepthNode();
-                }
-            }, false, RootTag.class);
-            Class<?>[] syntaxTags = new Class<?>[this.sourceElements.size()];
+            Class<?>[] syntaxTags = new Class<?>[this.sourceElements.size() + (hasRootElement ? 0 : 1)];
             Iterator<SourceElement> elementsIterator = this.sourceElements.iterator();
-            for (int i = 0; i < syntaxTags.length; i++) {
-                syntaxTags[i] = elementsIterator.next().getTag();
+            int i = 0;
+            while (elementsIterator.hasNext()) {
+                syntaxTags[i++] = elementsIterator.next().getTag();
+            }
+            assert i == sourceElements.size();
+            if (!hasRootElement) {
+                syntaxTags[i] = RootTag.class;
             }
             this.syntaxElementsBinding = createBinding(includeInternalCode, sFilter, new ExecutionEventNodeFactory() {
                 @Override
                 public ExecutionEventNode create(EventContext context) {
-                    return new SteppingNode(context);
+                    if (context.hasTag(RootTag.class)) {
+                        return new RootSteppingDepthNode(context);
+                    } else {
+                        return new SteppingNode(context);
+                    }
                 }
             }, hasExpressionElement, syntaxTags);
-            this.callBinding = createBinding(includeInternalCode, sFilter, new ExecutionEventNodeFactory() {
-                @Override
-                public ExecutionEventNode create(EventContext context) {
-                    return new CallSteppingNode(context);
-                }
-            }, false, CallTag.class);
             allBindings.add(syntaxElementsBinding);
-            allBindings.add(callBinding);
         }
     }
 
@@ -545,13 +537,8 @@ public final class DebuggerSession implements Closeable {
     private void removeBindings() {
         assert Thread.holdsLock(this);
         if (syntaxElementsBinding != null) {
-            allBindings.remove(callBinding);
             allBindings.remove(syntaxElementsBinding);
-            callBinding.dispose();
             syntaxElementsBinding.dispose();
-            callBinding = null;
-            rootBinding.dispose();
-            rootBinding = null;
             syntaxElementsBinding = null;
             if (Debugger.TRACE) {
                 trace("disabled stepping");
@@ -835,52 +822,14 @@ public final class DebuggerSession implements Closeable {
             breakpointFailures.put(fb, conditionFailure.getConditionFailure());
         }
 
-        List<Breakpoint> breaks = null;
-        for (DebuggerNode node : nodes) {
-            Breakpoint breakpoint = node.getBreakpoint();
-            if (breakpoint == null || !isBreakpointsActive(breakpoint.getKind())) {
-                continue; // not a breakpoint node
-            }
-            boolean hit = true;
-            BreakpointConditionFailure failure = null;
-            try {
-                hit = breakpoint.notifyIndirectHit(source, node, frame, exception);
-            } catch (BreakpointConditionFailure e) {
-                failure = e;
-            }
-            if (hit) {
-                if (breaks == null) {
-                    breaks = new ArrayList<>();
-                }
-                breaks.add(breakpoint.isGlobal() ? breakpoint.getROWrapper() : breakpoint);
-            }
-            if (failure != null) {
-                if (breakpointFailures == null) {
-                    breakpointFailures = new HashMap<>();
-                }
-                Breakpoint fb = failure.getBreakpoint();
-                if (fb.isGlobal()) {
-                    fb = fb.getROWrapper();
-                }
-                breakpointFailures.put(fb, failure.getConditionFailure());
-            }
-        }
-
-        boolean hitStepping = s.step(this, source.getContext(), suspendAnchor);
-        boolean hitBreakpoint = breaks != null && !breaks.isEmpty();
-        Object newReturnValue = returnValue;
-        if (hitStepping || hitBreakpoint) {
-            s.consume();
-            newReturnValue = doSuspend(SuspendedContext.create(source.getContext(), debugger.getEnv()), suspendAnchor, frame, source, inputValuesProvider, returnValue, exception, breaks,
-                            breakpointFailures);
-        } else {
-            if (Debugger.TRACE) {
-                trace("ignored suspended reason: strategy(%s) from source:%s context:%s location:%s", s, source, source.getContext(), source.getSuspendAnchors());
-            }
-        }
-        if (s.isKill()) {   // ComposedStrategy can become kill
-            throw new KillException(source.getContext().getInstrumentedNode());
-        }
+        Object newReturnValue = processBreakpointsAndStep(nodes, s, source, frame, suspendAnchor,
+                        inputValuesProvider, returnValue, exception, breakpointFailures,
+                        new Supplier<SuspendedContext>() {
+                            @Override
+                            public SuspendedContext get() {
+                                return SuspendedContext.create(source.getContext(), debugger.getEnv());
+                            }
+                        });
         return newReturnValue;
     }
 
@@ -944,6 +893,125 @@ public final class DebuggerSession implements Closeable {
         doSuspend(context, SuspendAnchor.AFTER, caller.frame, insertableNode, null, null, null, Collections.emptyList(), Collections.emptyMap());
     }
 
+    private Object notifyCallerReturn(SteppingStrategy s, DebuggerNode source, SuspendAnchor suspendAnchor, Object returnValue) {
+        // SuspensionFilter:
+        if (source.isStepNode()) {
+            if (ignoreLanguageContextInitialization.get() && !source.getContext().isLanguageContextInitialized()) {
+                return returnValue;
+            }
+        }
+        // Fake the caller context
+        class Caller {
+            final Node node;
+            final MaterializedFrame frame;
+
+            Caller(FrameInstance frameInstance, Node callNode) {
+                this.node = callNode;
+                this.frame = frameInstance.getFrame(FrameAccess.MATERIALIZE).materialize();
+            }
+        }
+        Caller caller = Truffle.getRuntime().iterateFrames(new FrameInstanceVisitor<Caller>() {
+            private int depth = 0;
+
+            @Override
+            public Caller visitFrame(FrameInstance frameInstance) {
+                // we stop at eval root stack frames
+                if (!SuspendedEvent.isEvalRootStackFrame(DebuggerSession.this, frameInstance) && (depth++ == 0)) {
+                    return null;
+                }
+                Node callNode = frameInstance.getCallNode();
+                while (callNode != null && !SourceSectionFilter.ANY.includes(callNode)) {
+                    callNode = callNode.getParent();
+                }
+                if (callNode == null) {
+                    return null;
+                }
+                RootNode root = callNode.getRootNode();
+                if (root == null || !includeInternal && root.isInternal()) {
+                    return null;
+                }
+                return new Caller(frameInstance, callNode);
+            }
+        });
+        if (caller == null) {
+            // We did not find a caller node
+            return returnValue;
+        }
+        List<DebuggerNode> nodes = collectDebuggerNodes(caller.node, suspendAnchor);
+        for (DebuggerNode node : nodes) {
+            Breakpoint breakpoint = node.getBreakpoint();
+            if (breakpoint == null || isBreakpointsActive(breakpoint.getKind()) && breakpoint.getCondition() == null) {
+                // Not a breakpoint node, nor unconditional breakpoint.
+                // We will suspend there later on.
+                return returnValue;
+            }
+        }
+        // Suspend on the return from the caller and mark all existing nodes as duplicate
+        for (DebuggerNode node : nodes) {
+            node.markAsDuplicate(this);
+        }
+        Object newReturnValue = processBreakpointsAndStep(nodes, s, source, caller.frame, suspendAnchor, null, returnValue, null, null, new Supplier<SuspendedContext>() {
+            @Override
+            public SuspendedContext get() {
+                return SuspendedContext.create(caller.node, null);
+            }
+        });
+        return newReturnValue;
+    }
+
+    @SuppressWarnings("all") // The parameter breakpointFailures should not be assigned
+    private Object processBreakpointsAndStep(List<DebuggerNode> nodes, SteppingStrategy s, DebuggerNode source, MaterializedFrame frame,
+                    SuspendAnchor suspendAnchor, InputValuesProvider inputValuesProvider, Object returnValue, DebugException exception,
+                    Map<Breakpoint, Throwable> breakpointFailures, Supplier<SuspendedContext> contextSupplier) {
+        List<Breakpoint> breaks = null;
+        for (DebuggerNode node : nodes) {
+            Breakpoint breakpoint = node.getBreakpoint();
+            if (breakpoint == null || !isBreakpointsActive(breakpoint.getKind())) {
+                continue; // not a breakpoint node
+            }
+            boolean hit = true;
+            BreakpointConditionFailure failure = null;
+            try {
+                hit = breakpoint.notifyIndirectHit(source, node, frame, exception);
+            } catch (BreakpointConditionFailure e) {
+                failure = e;
+            }
+            if (hit) {
+                if (breaks == null) {
+                    breaks = new ArrayList<>();
+                }
+                breaks.add(breakpoint.isGlobal() ? breakpoint.getROWrapper() : breakpoint);
+            }
+            if (failure != null) {
+                if (breakpointFailures == null) {
+                    breakpointFailures = new HashMap<>();
+                }
+                Breakpoint fb = failure.getBreakpoint();
+                if (fb.isGlobal()) {
+                    fb = fb.getROWrapper();
+                }
+                breakpointFailures.put(fb, failure.getConditionFailure());
+            }
+        }
+
+        boolean hitStepping = s.step(this, source.getContext(), suspendAnchor);
+        boolean hitBreakpoint = breaks != null && !breaks.isEmpty();
+        Object newReturnValue = returnValue;
+        if (hitStepping || hitBreakpoint) {
+            s.consume();
+            newReturnValue = doSuspend(contextSupplier.get(), suspendAnchor, frame, source, inputValuesProvider, returnValue, exception, breaks,
+                            breakpointFailures);
+        } else {
+            if (Debugger.TRACE) {
+                trace("ignored suspended reason: strategy(%s) from source:%s context:%s location:%s", s, source, source.getContext(), source.getSuspendAnchors());
+            }
+        }
+        if (s.isKill()) {   // ComposedStrategy can become kill
+            throw new KillException(source.getContext().getInstrumentedNode());
+        }
+        return newReturnValue;
+    }
+
     private Object doSuspend(SuspendedContext context, SuspendAnchor suspendAnchor, MaterializedFrame frame,
                     InsertableNode insertableNode, InputValuesProvider inputValuesProvider, Object returnValue, DebugException exception,
                     List<Breakpoint> breaks, Map<Breakpoint, Throwable> conditionFailures) {
@@ -997,7 +1065,7 @@ public final class DebuggerSession implements Closeable {
         if (strategy.isKill()) {
             throw new KillException(context.getInstrumentedNode());
         } else if (strategy.isUnwind()) {
-            ThreadDeath unwind = context.createUnwind(null, rootBinding);
+            ThreadDeath unwind = context.createUnwind(null, syntaxElementsBinding);
             ((SteppingStrategy.Unwind) strategy).unwind = unwind;
             throw unwind;
         }
@@ -1032,6 +1100,17 @@ public final class DebuggerSession implements Closeable {
                 if (node.isActiveAt(suspendAnchor)) {
                     nodes.add(node);
                 }
+            }
+        }
+        return nodes;
+    }
+
+    private List<DebuggerNode> collectDebuggerNodes(Node iNode, SuspendAnchor suspendAnchor) {
+        List<DebuggerNode> nodes = new ArrayList<>();
+        for (EventBinding<?> binding : allBindings) {
+            DebuggerNode node = (DebuggerNode) debugger.getInstrumenter().lookupExecutionEventNode(iNode, binding);
+            if (node != null && node.isActiveAt(suspendAnchor)) {
+                nodes.add(node);
             }
         }
         return nodes;
@@ -1124,7 +1203,7 @@ public final class DebuggerSession implements Closeable {
         }
     }
 
-    private final class SteppingNode extends DebuggerNode implements InputValuesProvider {
+    private class SteppingNode extends DebuggerNode implements InputValuesProvider {
 
         SteppingNode(EventContext context) {
             super(context);
@@ -1185,7 +1264,7 @@ public final class DebuggerSession implements Closeable {
         }
 
         @TruffleBoundary
-        private Object doStepAfter(MaterializedFrame frame, Object result) {
+        protected final Object doStepAfter(MaterializedFrame frame, Object result) {
             SuspendAnchor anchor = SuspendAnchor.AFTER;
             SteppingStrategy steppingStrategy = getSteppingStrategy(Thread.currentThread());
             if (steppingStrategy != null && steppingStrategy.isActiveOnStepTo(context, anchor)) {
@@ -1216,20 +1295,28 @@ public final class DebuggerSession implements Closeable {
 
     }
 
-    private final class CallSteppingNode extends DebuggerNode {
+    /**
+     * Combines stepping with stack depth control, stop after a call and unwind.
+     */
+    private final class RootSteppingDepthNode extends SteppingNode {
 
-        CallSteppingNode(EventContext context) {
+        RootSteppingDepthNode(EventContext context) {
             super(context);
         }
 
         @Override
-        EventBinding<?> getBinding() {
-            return callBinding;
+        boolean isStepNode() {
+            return hasRootElement;
         }
 
         @Override
-        boolean isStepNode() {
-            return true;
+        protected void onEnter(VirtualFrame frame) {
+            if (stepping.get()) {
+                doEnter();
+                if (hasRootElement) {
+                    super.onEnter(frame);
+                }
+            }
         }
 
         @Override
@@ -1242,59 +1329,19 @@ public final class DebuggerSession implements Closeable {
         @Override
         public void onReturnExceptional(VirtualFrame frame, Throwable exception) {
             if (stepping.get()) {
-                doReturn(frame.materialize(), null);
-            }
-        }
-
-        @TruffleBoundary
-        private void doReturn(MaterializedFrame frame, Object result) {
-            SteppingStrategy steppingStrategy = strategyMap.get(Thread.currentThread());
-            if (steppingStrategy != null && steppingStrategy.isStopAfterCall()) {
-                Object newResult = notifyCallback(this, frame, SuspendAnchor.AFTER, null, result, null, null);
-                if (newResult != result) {
-                    CompilerDirectives.transferToInterpreter();
-                    throw getContext().createUnwind(new ChangedReturnInfo(newResult));
+                if (hasRootElement) {
+                    doStepAfter(frame.materialize(), exception);
                 }
-            }
-        }
-
-        @Override
-        Set<SuspendAnchor> getSuspendAnchors() {
-            return DebuggerSession.ANCHOR_SET_AFTER;
-        }
-
-        @Override
-        boolean isActiveAt(SuspendAnchor anchor) {
-            return SuspendAnchor.AFTER == anchor;
-        }
-
-    }
-
-    private final class RootSteppingDepthNode extends ExecutionEventNode implements InsertableNode {
-
-        @Override
-        protected void onEnter(VirtualFrame frame) {
-            if (stepping.get()) {
-                doEnter();
-            }
-        }
-
-        @Override
-        public void onReturnValue(VirtualFrame frame, Object result) {
-            if (stepping.get()) {
-                doReturn();
-            }
-        }
-
-        @Override
-        public void onReturnExceptional(VirtualFrame frame, Throwable exception) {
-            if (stepping.get()) {
                 doReturn();
             }
         }
 
         @Override
         protected Object onUnwind(VirtualFrame frame, Object info) {
+            Object ret = super.onUnwind(frame, info);
+            if (ret != null) {
+                return ret;
+            }
             if (stepping.get()) {
                 return doUnwind(frame.materialize());
             } else {
@@ -1312,6 +1359,28 @@ public final class DebuggerSession implements Closeable {
             SteppingStrategy steppingStrategy = strategyMap.get(Thread.currentThread());
             if (steppingStrategy != null) {
                 steppingStrategy.notifyCallEntry();
+            }
+        }
+
+        @TruffleBoundary
+        private void doReturn(MaterializedFrame frame, Object result) {
+            SteppingStrategy steppingStrategy;
+            try {
+                if (hasRootElement) {
+                    doStepAfter(frame, result);
+                }
+            } finally {
+                steppingStrategy = strategyMap.get(Thread.currentThread());
+                if (steppingStrategy != null) {
+                    // Stepping out of a function.
+                    steppingStrategy.notifyCallExit();
+                }
+            }
+            if (steppingStrategy != null && steppingStrategy.isStopAfterCall()) {
+                Object newResult = notifyCallerReturn(steppingStrategy, this, SuspendAnchor.AFTER, result);
+                if (newResult != result) {
+                    throw getContext().createUnwind(new ChangedReturnInfo(newResult));
+                }
             }
         }
 
@@ -1336,6 +1405,17 @@ public final class DebuggerSession implements Closeable {
                 return null;
             }
         }
+
+        @Override
+        Set<SuspendAnchor> getSuspendAnchors() {
+            return DebuggerSession.ANCHOR_SET_ALL;
+        }
+
+        @Override
+        boolean isActiveAt(SuspendAnchor anchor) {
+            return hasRootElement;
+        }
+
     }
 
     /**